Clean up files and folders to be more organized
