+++ client/src/components/NotesPage/AddEditNoteDialog.module.css
@@ -1,0 +1,15 @@
+.blockCenter {
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+.flexCenter {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.width100 {
+    width: 100%;
+}
