+++ client/src/components/NotesPage/Note.module.css
@@ -1,0 +1,23 @@
+.noteCard {
+    background-color: red;
+}
+
+.cardTypography {
+    color: blue;
+}
+
+.blockCenter {
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+.flexCenter {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.width100 {
+    width: 100%;
+}
