--- client/src/styles/Note.module.css
@@ -1,7 +1,0 @@
-.noteCard {
-    background-color: red;
-}
-
-.cardTypography {
-    color: blue;
-}
