--- client/src/styles/utils.module.css
@@ -1,15 +1,0 @@
-.blockCenter {
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.flexCenter {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.width100 {
-    width: 100%;
-}
