--- server/src/util/asserIsDefined.ts
@@ -1,5 +1,0 @@
-export function asserIsDefined<T>(val: T): asserts val is NonNullable<T> {
-  if (!val) {
-    throw Error("Expected 'val' to be defined, but received " + val);
-  }
-}
